# SharpAI.Client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "SharpAI.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/systemprompt/set": {
      "post": {
        "tags": [
          "Llama"
        ],
        "parameters": [
          {
            "name": "updateCurrentContext",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/model/load": {
      "post": {
        "tags": [
          "Llama"
        ],
        "parameters": [
          {
            "name": "fuzzyMatch",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "forceUnload",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LlamaModelLoadRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LlamaModelLoadRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LlamaModelLoadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LlamaModelFile"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlamaModelFile"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlamaModelFile"
                }
              }
            }
          }
        }
      }
    },
    "/model/load/simple": {
      "post": {
        "tags": [
          "Llama"
        ],
        "parameters": [
          {
            "name": "modelNameOrPath",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fuzzyMatch",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "forceUnload",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "tryMultimodal",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "maxTokens",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1024
            }
          },
          {
            "name": "backend",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/LlamaBackend"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LlamaModelFile"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlamaModelFile"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlamaModelFile"
                }
              }
            }
          }
        }
      }
    },
...
    "/context/list": {
      "get": {
        "tags": [
          "Llama"
        ],
        "parameters": [
          {
            "name": "sortedByLatest",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LlamaContextData"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LlamaContextData"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LlamaContextData"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/context/create": {
      "post": {
        "tags": [
          "Llama"
        ],
        "parameters": [
          {
            "name": "use",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "saveJson",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/context/save": {
      "post": {
        "tags": [
          "Llama"
        ],
        "parameters": [
          {
            "name": "differentNameOrPath",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/context/load": {
      "post": {
        "tags": [
          "Llama"
        ],
        "parameters": [
          {
            "name": "contextNameOrPath",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LlamaContextData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlamaContextData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlamaContextData"
                }
              }
            }
          }
        }
      }
    },
    "/context/rename": {
      "post": {
        "tags": [
          "Llama"
        ],
        "parameters": [
          {
            "name": "contextNameOrPath",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "newName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/context/delete": {
      "delete": {
        "tags": [
          "Llama"
        ],
        "parameters": [
          {
            "name": "contextNameOrPath",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean",
                  "nullable": true
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "nullable": true
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean",
                  "nullable": true
                }
              }
            }
          }
        }
      }
    },
    "/generate": {
      "post": {
        "tags": [
          "Llama"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LlamaGenerationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LlamaGenerationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LlamaGenerationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/generate/simple": {
      "post": {
        "tags": [
          "Llama"
        ],
        "parameters": [
          {
            "name": "prompt",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxTokens",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1024
            }
          },
          {
            "name": "temperature",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "float",
              "default": 0.7
            }
          },
          {
            "name": "topP",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "float",
              "default": 0.95
            }
          },
          {
            "name": "useSystemPrompt",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/generate/text": {
      "post": {
        "tags": [
          "Llama"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LlamaGenerationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LlamaGenerationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LlamaGenerationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LlamaBackend": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "LlamaModelFile": {
        "type": "object",
        "properties": {
          "filePath": {
            "type": "string",
            "nullable": true
          },
          "modelName": {
            "type": "string",
            "nullable": true
          },
          "modelDescriptor": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "fileSizeMb": {
            "type": "number",
            "format": "double"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "mmProjFilePath": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LlamaModelLoadRequest": {
        "type": "object",
        "properties": {
          "modelFile": {
            "$ref": "#/components/schemas/LlamaModelFile"
          },
          "mmProjPath": {
            "type": "string",
            "nullable": true
          },
          "contextSize": {
            "type": "integer",
            "format": "int32"
          },
          "backend": {
            "$ref": "#/components/schemas/LlamaBackend"
          },
          "tryMultimodal": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}